requirements.txt
flask>=3.0
python-pptx>=0.6.23
Pillow>=10.0
